refactor(arch): move word.py to bytecode/ with re-export shim (stage 2 of #569)
Second stage of the multi-stage package merge tracked in #569. Moves
the canonical home of Word from bloqade.lanes.layout.word to
bloqade.lanes.bytecode.word — foundation types live alongside the
Rust _native bindings they wrap.

Old import path bloqade.lanes.layout.word remains as a thin
re-export shim, so every existing caller across bloqade.lanes,
bloqade.gemini, demos, and tests compiles unchanged. The shim is
removed in the final cleanup stage (Stage 9 of the plan) once all
in-flight branches have rebased onto the canonical path.

bloqade/lanes/bytecode/__init__.py is left alone for the same
reason as Stage 1: switching its Word re-export source from _native
to .word would change the class object callers receive (raw Rust
class -> Python wrapper) — a real semantic break. Deferred to
Stage 9.

Refs #569.
